pyproject.toml
[project]
name = "monsterui-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.40.0",
    "boto3>=1.35.82",
    "claudette>=0.1.0",
    "lxml>=5.3.0",
    "mistletoe>=1.4.0",
    "monsterui>=0.0.6",
    "python-fasthtml>=0.10.1",
]
